+++ packages/okta-react/src/AuthService.js
@@ -190,15 +190,24 @@ class AuthService {
   }
 
   async login(fromUri, additionalParams) {
-    // Set pending state to avoid concurrent login flows
-    const authState = this.getAuthState();
-    this.emitAuthState({ ...authState, isPending: true });
+    if(this._pending.handleLogin) { 
+      // Don't trigger second round
+      return;
+    }
+
+    this._pending.handleLogin = true;
+    // Update UI pending state
+    this.emitAuthState({ ...this.getAuthState(), isPending: true });
     // Save the current url before redirect
     this.setFromUri(fromUri); // will save current location if fromUri is undefined
-    if (this._config.onAuthRequired) {
-      return this._config.onAuthRequired(this);
+    try {
+      if (this._config.onAuthRequired) {
+        return await this._config.onAuthRequired(this);
+      }
+      return await this.redirect(additionalParams);
+    } finally {
+      this._pending.handleLogin = null;
     }
-    return this.redirect(additionalParams);
   }
 
   _convertLogoutPathToOptions(redirectUri) {