+++ packages/okta-react/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.0.4
+
+### Bug Fixes
+
+- [#848](https://github.com/okta/okta-oidc-js/pull/848) Removes `onSessionExpired` behavior.
+
 # 3.0.3
 
 ### Bug Fixes